requirements.txt
google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-cloud==0.32.0
slackclient
docker==3.2.0